re-organize some docs
